chore(tests): move ExtractRemainingFromPayload to plugin_collection
