+++ falco_plugin_tests/src/plugin_collection/extract/mod.rs
@@ -1,4 +1,5 @@
 pub mod extra_fields;
 pub mod nested;
+pub mod remaining_from_payload;
 pub mod remaining_from_table;
 pub mod remaining_from_table_runtime;
